add academy days section, tweak styles
