+++ styles/pages/Academy.module.scss
@@ -18,7 +18,7 @@
 }
 
 .mainContentContainer {
-    margin: 0 auto $grid * 3 auto;
+    margin: 0 auto $grid * 2 auto;
     max-width: $content--narrow + 29px;
     padding: 0 $grid * 1.5;
 
@@ -27,7 +27,7 @@
     }
 
     @include media-query(large) {
-        margin: 0 auto $grid * 4 18%;
+        margin: 0 auto $grid * 3 18%;
         padding: 0;
     }
 }
